Add boilerplate thunk
