+++ .gitignore
@@ -15,7 +15,7 @@
 .env.development.local
 .env.test.local
 .env.production.local
-/src/Utils/APIKey.js
+/src/utils/APIKey.js
 
 npm-debug.log*
 yarn-debug.log*
