+++ package.json
@@ -11,6 +11,7 @@
     "react-scripts": "2.0.5",
     "redux": "^4.0.1",
     "redux-devtools-extension": "^2.13.5",
+    "redux-thunk": "^2.3.0",
     "uuid": "^3.3.2"
   },
   "scripts": {
