+++ src/Components/TitleCard/TitleCard.css
@@ -3,8 +3,8 @@
 }
 
 .poster {
-  height: 400px;
-  width: 300px;
+  height: 200px;
+  
 }
 
 .title {
