+++ src/Utils/API.js
@@ -1,0 +1,9 @@
+export const fetchData = async (url) => {
+  try {
+    const response = await fetch(url)
+    const data = await response.json()
+    return data
+  } catch (error) {
+    throw new Error(error.message)
+  }
+}
