+++ src/actions/index.js
@@ -1,3 +1,8 @@
+export const FetchMovieUser = ()
+
+
+
+
 export const addMovies = (movies) => ({
   type: "ADD_MOVIES",
   movies
