+++ src/components/TitleCard/index.js
@@ -5,7 +5,7 @@ export const TitleCard = (props) => {
   return(
     <div className='title-card'>
       <h2 className='title'>{props.title}</h2>
-      <img className='poster' src={`https://image.tmdb.org/t/p/w500/${props.image}`} />
+      <img className='poster' src={`https://image.tmdb.org/t/p/w500/${props.background}`} />
     </div>
 
   )
