+++ src/index.js
@@ -2,14 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { createStore } from 'redux';
-import { rootReducer } from './reducers';
+import { store } from './store';
+
 import './index.css';
 import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 
-const devTools = window.__REDUX_DEVTOOLS_EXTENSION__ / window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__;
-const store = createStore(rootReducer, devTools);
+
 
 
 const router = (
